Seperate out schema serialization from s2n-tls-metrics-subscriber
