+++ bindings/rust/standard/Cargo.toml
@@ -1,12 +1,16 @@
 [workspace]
 members = [
     "benchmarks",
-    "integration", 
+    "integration",
     "s2n-tls-hyper",
+    "s2n-tls-metrics-schema",
     "s2n-tls-metrics-subscriber",
     "s2n-tls-sys-internal",
     "tls-harness",
 ]
+
+[workspace.dependencies]
+s2n-codec = "0.80"
 
 # brass-aphid-wire has a dependency on s2n-tls. This patch is necessary to unify
 # the crates.io dependency that brass-aphid-wire declares with the path-based
