+++ bindings/rust/standard/s2n-tls-metrics-schema/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "s2n-tls-metrics-schema"
+version = "0.0.1"
+edition = "2024"
+description = "Serialization schema types for s2n-tls metrics. No stability guarantees."
+authors = ["AWS s2n"]
+repository = "https://github.com/aws/s2n-tls"
+license = "Apache-2.0"
+
+[dependencies]
+s2n-codec.workspace = true
+zerocopy = "0.8.33"
+serde = { version = "1.0.225", features = ["derive"] }
+serde_with = "3.16"
+tracing = "0.1.43"
+metrique-writer = { version = "0.1.19", optional = true }
+
+[features]
+default = []
+metrique = ["dep:metrique-writer"]
+
+[dev-dependencies]
+s2n-tls-sys-internal = { path = "../s2n-tls-sys-internal" }
+serde_json = "1.0.141"
+ciborium = "0.2"
