+++ bindings/rust/standard/s2n-tls-metrics-subscriber/src/record.rs
@@ -6,67 +6,39 @@ use std::{
     time::SystemTime,
 };
 
-use serde::{Deserialize, Serialize};
-
-use crate::{
-    attribution::Attribution,
-    compatibility::{Cnsa1, Cnsa2, Fips20251201, General20251201, TlsProfile},
-    counter::{Counter, FrozenCounter},
-    detector::SyntheticTrafficDetector,
-    label::{State, metric_label},
-    parsing::ClientHelloSupportedParameters,
+use s2n_tls_metrics_schema::{
+    record::FrozenHandshakeRecord,
     static_lists::{
-        CIPHER_COUNT, Cipher, FiniteCounter, GROUP_COUNT, Group, PROTOCOL_COUNT, SIGNATURE_COUNT,
-        Signature, TlsParam, ToStaticString, Version,
+        CIPHER_COUNT, Cipher, GROUP_COUNT, Group, PROTOCOL_COUNT, SIGNATURE_COUNT, Signature,
+        Version,
     },
 };
 
-/// Metric Record is an opaque type which implements [`metrique_writer::Entry`].
-///
-/// This is the preferred type for public s2n-tls-metric-subscriber traits and
-/// interfaces.
-// This currently just holds a single struct. In the future we will
-// likely rely on an enum to handle different record types, e.g. SessionResumptionFailure.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct MetricRecord {
-    pub attribution: Attribution,
-    pub handshake: FrozenHandshakeRecord,
-}
+use crate::{
+    compatibility::{Cnsa1, Cnsa2, Fips20251201, General20251201, TlsProfile},
+    counter::Counter,
+    detector::SyntheticTrafficDetector,
+    parsing::ClientHelloSupportedParameters,
+};
 
-impl MetricRecord {
-    pub(crate) fn new(handshake: FrozenHandshakeRecord, attribution: Attribution) -> Self {
-        Self {
-            attribution,
-            handshake,
-        }
-    }
-}
-
-impl metrique_writer::Entry for MetricRecord {
-    fn write<'a>(&'a self, writer: &mut impl metrique_writer::EntryWriter<'a>) {
-        // "Operation" is a reserved label in some telemetry systems,
-        // so the key generally shouldn't be changed.
-        let operation = if self.attribution.component.is_empty() {
-            "TlsTelemetry".to_owned()
-        } else {
-            format!("TlsTelemetry.{}", self.attribution.component)
-        };
-        writer.value("Operation", &operation);
-        writer.value("service", &self.attribution.service);
-        writer.value("resource", &self.attribution.resource);
-        self.handshake.write(writer)
-    }
+fn protocol_version_to_iana(v: s2n_tls::enums::Version) -> Option<Version> {
+    let iana = match v {
+        s2n_tls::enums::Version::SSLV3 => 0x0300u16,
+        s2n_tls::enums::Version::TLS10 => 0x0301,
+        s2n_tls::enums::Version::TLS11 => 0x0302,
+        s2n_tls::enums::Version::TLS12 => 0x0303,
+        s2n_tls::enums::Version::TLS13 => 0x0304,
+        _ => return None,
+    };
+    Some(Version(s2n_codec::zerocopy::U16::new(iana)))
 }
 
 /// The HandshakeRecordInProgress stores the in-flight counters for handshake
 /// information - e.g. negotiated parameters.
 #[derive(Debug)]
 pub(crate) struct HandshakeRecordInProgress {
-    /// This is used to send a frozen version back to the Aggregator, after which
-    /// point it can be exported. This is only used in the drop impl.
     exporter: std::sync::mpsc::Sender<FrozenHandshakeRecord>,
 
-    /// the total number of handshakes that this record represents.
     handshake_count: AtomicU64,
 
     negotiated_protocols: Counter<PROTOCOL_COUNT, Version>,
@@ -74,8 +46,6 @@ pub(crate) struct HandshakeRecordInProgress {
     negotiated_groups: Counter<GROUP_COUNT, Group>,
     negotiated_signatures: Counter<SIGNATURE_COUNT, Signature>,
 
-    // we do not attempt to detect supported parameters for SSLv2 formatted client
-    // hellos
     sslv2_client_hello: AtomicU64,
     supported_protocols: Counter<PROTOCOL_COUNT, Version>,
     supported_ciphers: Counter<CIPHER_COUNT, Cipher>,
@@ -87,18 +57,9 @@ pub(crate) struct HandshakeRecordInProgress {
     compatibility_cnsa1: AtomicU64,
     compatibility_cnsa2: AtomicU64,
 
-    /// sum of handshake duration, including network latency and waiting
-    ///
-    /// To get the average, divide this by handshake_count.
     handshake_duration_us: AtomicU64,
-    /// sum of handshake compute
-    ///
-    /// To get the average, divide this by handshake_count.
     handshake_compute_us: AtomicU64,
 
-    /// Number of handshakes flagged by the configured
-    /// [`SyntheticTrafficDetector`]. Synthetic handshakes are excluded from
-    /// every other counter on this record (including `handshake_count`).
     synthetic_traffic_count: AtomicU64,
 }
 
@@ -138,11 +99,6 @@ impl HandshakeRecordInProgress {
     ) -> Result<(), Box<dyn std::error::Error>> {
         let client_hello = conn.client_hello()?;
 
-        // Run the detector first, on every handshake (including SSLv2).
-        // Synthetic handshakes contribute ONLY to `synthetic_traffic_count`;
-        // every other counter (including `handshake_count`) reflects real
-        // traffic only, so consumers can read each metric directly without
-        // post-processing.
         if let Some(detector) = detector {
             if detector.is_synthetic(client_hello) {
                 self.synthetic_traffic_count.fetch_add(1, Ordering::Relaxed);
@@ -151,10 +107,6 @@ impl HandshakeRecordInProgress {
         }
 
         self.handshake_count.fetch_add(1, Ordering::Relaxed);
-
-        ////////////////////////////////////////////////////////////////////////
-        /////////////////////   fields from connection   ///////////////////////
-        ////////////////////////////////////////////////////////////////////////
 
         if let Some(sig) = conn.signature_scheme().and_then(|s| s.parse().ok()) {
             self.negotiated_signatures.increment(&sig);
@@ -203,11 +155,7 @@ impl HandshakeRecordInProgress {
             }
         }
 
-        ////////////////////////////////////////////////////////////////////////
-        //////////////////////   fields from event   ///////////////////////////
-        ////////////////////////////////////////////////////////////////////////
-
-        if let Ok(version) = event.protocol_version().to_static_string().parse() {
+        if let Some(version) = protocol_version_to_iana(event.protocol_version()) {
             self.negotiated_protocols.increment(&version);
         }
 
@@ -219,9 +167,6 @@ impl HandshakeRecordInProgress {
             self.negotiated_groups.increment(&group);
         }
 
-        // accuracy: as long as the handshake took less than 500,000 years
-        // this cast will not truncate. We prefer truncation/less accurate metrics
-        // over a panic.
         self.handshake_compute_us.fetch_add(
             event.synchronous_time().as_micros() as u64,
             Ordering::Relaxed,
@@ -232,16 +177,6 @@ impl HandshakeRecordInProgress {
         Ok(())
     }
 
-    /// make a copy of this record to be exported.
-    ///
-    /// ### A Note On Ordering Correctness
-    ///
-    /// It is important that this function observes the results of all the `fetch_add`
-    /// operations on other threads.
-    ///
-    /// Simple Intuition: This function takes a `&mut`. Therefore the rust compiler
-    /// enforces that there are no other references to this memory and there isn't
-    /// anything to actually synchronize. So a Relaxed load is fine.
     fn finish(&mut self) -> FrozenHandshakeRecord {
         FrozenHandshakeRecord {
             freeze_time: SystemTime::now(),
@@ -274,185 +209,19 @@ impl HandshakeRecordInProgress {
 impl Drop for HandshakeRecordInProgress {
     fn drop(&mut self) {
         let frozen = self.finish();
-        // no available way to report error
         let _ = self.exporter.send(frozen);
-    }
-}
-
-/// `SystemTime` has no meaningful `Default` impl for serde's purposes, so we
-/// pick `UNIX_EPOCH` explicitly as the `#[serde(default = ...)]` target for
-/// `freeze_time`.
-fn system_time_epoch() -> SystemTime {
-    SystemTime::UNIX_EPOCH
-}
-
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct FrozenHandshakeRecord {
-    #[serde(default = "system_time_epoch")]
-    pub freeze_time: SystemTime,
-
-    #[serde(default)]
-    pub handshake_count: u64,
-
-    #[serde(default)]
-    pub negotiated_protocols: FrozenCounter<PROTOCOL_COUNT, Version>,
-    #[serde(default)]
-    pub negotiated_ciphers: FrozenCounter<CIPHER_COUNT, Cipher>,
-    #[serde(default)]
-    pub negotiated_groups: FrozenCounter<GROUP_COUNT, Group>,
-    #[serde(default)]
-    pub negotiated_signatures: FrozenCounter<SIGNATURE_COUNT, Signature>,
-
-    #[serde(default)]
-    pub sslv2_client_hello: u64,
-    #[serde(default)]
-    pub supported_protocols: FrozenCounter<PROTOCOL_COUNT, Version>,
-    #[serde(default)]
-    pub supported_ciphers: FrozenCounter<CIPHER_COUNT, Cipher>,
-    #[serde(default)]
-    pub supported_groups: FrozenCounter<GROUP_COUNT, Group>,
-    #[serde(default)]
-    pub supported_signatures: FrozenCounter<SIGNATURE_COUNT, Signature>,
-
-    #[serde(default)]
-    pub compatibility_general20251201: u64,
-    #[serde(default)]
-    pub compatibility_fips20251201: u64,
-    #[serde(default)]
-    pub compatibility_cnsa1: u64,
-    #[serde(default)]
-    pub compatibility_cnsa2: u64,
-
-    #[serde(default)]
-    pub handshake_duration_us: u64,
-    #[serde(default)]
-    pub handshake_compute_us: u64,
-
-    #[serde(default)]
-    pub synthetic_traffic_count: u64,
-}
-
-// This is just cfg(test) because we only use it in tests to assert on cases of
-// all-zero records
-#[cfg(test)]
-impl Default for FrozenHandshakeRecord {
-    fn default() -> Self {
-        Self {
-            freeze_time: SystemTime::UNIX_EPOCH,
-            handshake_count: 0,
-            negotiated_protocols: FrozenCounter::default(),
-            negotiated_ciphers: FrozenCounter::default(),
-            negotiated_groups: FrozenCounter::default(),
-            negotiated_signatures: FrozenCounter::default(),
-            sslv2_client_hello: 0,
-            supported_protocols: FrozenCounter::default(),
-            supported_ciphers: FrozenCounter::default(),
-            supported_groups: FrozenCounter::default(),
-            supported_signatures: FrozenCounter::default(),
-            compatibility_general20251201: 0,
-            compatibility_fips20251201: 0,
-            compatibility_cnsa1: 0,
-            compatibility_cnsa2: 0,
-            handshake_duration_us: 0,
-            handshake_compute_us: 0,
-            synthetic_traffic_count: 0,
-        }
-    }
-}
-
-impl metrique_writer::Entry for FrozenHandshakeRecord {
-    fn write<'a>(&'a self, writer: &mut impl metrique_writer::EntryWriter<'a>) {
-        writer.timestamp(self.freeze_time);
-
-        // Emit one label per non-zero slot for each (kind, state) cell.
-        // The label uses the element's `Display` impl.
-        fn write_counter<'a, const N: usize, T, W>(
-            counter: &'a FrozenCounter<N, T>,
-            parameter: TlsParam,
-            state: State,
-            writer: &mut W,
-        ) where
-            T: FiniteCounter<N> + std::fmt::Display,
-            W: metrique_writer::EntryWriter<'a>,
-        {
-            for (slot, element, count) in counter.iter_non_zero() {
-                let label = metric_label(slot, element, parameter, state);
-                writer.value(label, &count);
-            }
-        }
-
-        write_counter(
-            &self.negotiated_protocols,
-            TlsParam::Version,
-            State::Negotiated,
-            writer,
-        );
-        write_counter(
-            &self.negotiated_ciphers,
-            TlsParam::Cipher,
-            State::Negotiated,
-            writer,
-        );
-        write_counter(
-            &self.negotiated_groups,
-            TlsParam::Group,
-            State::Negotiated,
-            writer,
-        );
-        write_counter(
-            &self.negotiated_signatures,
-            TlsParam::SignatureScheme,
-            State::Negotiated,
-            writer,
-        );
-        write_counter(
-            &self.supported_protocols,
-            TlsParam::Version,
-            State::Supported,
-            writer,
-        );
-        write_counter(
-            &self.supported_ciphers,
-            TlsParam::Cipher,
-            State::Supported,
-            writer,
-        );
-        write_counter(
-            &self.supported_groups,
-            TlsParam::Group,
-            State::Supported,
-            writer,
-        );
-        write_counter(
-            &self.supported_signatures,
-            TlsParam::SignatureScheme,
-            State::Supported,
-            writer,
-        );
-
-        writer.value(
-            "compatibility.general20251201",
-            &self.compatibility_general20251201,
-        );
-        writer.value(
-            "compatibility.fips20251201",
-            &self.compatibility_fips20251201,
-        );
-        writer.value("compatibility.cnsa1", &self.compatibility_cnsa1);
-        writer.value("compatibility.cnsa2", &self.compatibility_cnsa2);
-
-        writer.value("sslv2_client_hello", &self.sslv2_client_hello);
-        writer.value("handshake_count", &self.handshake_count);
-        writer.value("handshake_duration_us", &self.handshake_duration_us);
-        writer.value("handshake_compute_us", &self.handshake_compute_us);
-        writer.value("synthetic_traffic_count", &self.synthetic_traffic_count);
     }
 }
 
 #[cfg(test)]
 mod tests {
-    use super::*;
     use crate::test_utils::{ARBITRARY_POLICY_1, TestEndpoint};
+    use s2n_tls_metrics_schema::{
+        counter::FrozenCounter,
+        static_lists::{Cipher, FiniteCounter},
+    };
+
+    use super::*;
 
     #[test]
     fn record_contents_negotiated_parameters() {
@@ -461,49 +230,65 @@ mod tests {
         let result = endpoint.client_handshake(&ARBITRARY_POLICY_1);
         endpoint.subscriber.finish_record();
         let records = endpoint.sink.records.lock().unwrap();
-        let record = &records[0].handshake;
+        let record = &records[0].as_schema().handshake;
 
         assert_eq!(record.handshake_count, 1);
-        assert_eq!(record.negotiated_ciphers.total(), 1);
-        assert_eq!(record.negotiated_groups.total(), 1);
-        assert_eq!(record.negotiated_signatures.total(), 1);
-        assert_eq!(record.negotiated_protocols.total(), 1);
-
-        let expected_version = result
-            .client
-            .actual_protocol_version()
-            .unwrap()
-            .to_static_string();
-        assert_eq!(record.negotiated_protocols.count_for(expected_version), 1);
-
-        let expected_cipher = result.client.cipher_suite().unwrap().to_owned();
-        let expected_cipher_description = Cipher::from_openssl_name(expected_cipher.as_str())
-            .and_then(|cipher| cipher.known_description())
-            .unwrap();
         assert_eq!(
             record
                 .negotiated_ciphers
-                .count_for(expected_cipher_description),
+                .iter_non_zero()
+                .map(|(_, _, c)| c)
+                .sum::<u64>(),
             1
         );
+        assert_eq!(
+            record
+                .negotiated_groups
+                .iter_non_zero()
+                .map(|(_, _, c)| c)
+                .sum::<u64>(),
+            1
+        );
+        assert_eq!(
+            record
+                .negotiated_signatures
+                .iter_non_zero()
+                .map(|(_, _, c)| c)
+                .sum::<u64>(),
+            1
+        );
+        assert_eq!(
+            record
+                .negotiated_protocols
+                .iter_non_zero()
+                .map(|(_, _, c)| c)
+                .sum::<u64>(),
+            1
+        );
+
+        let expected_version_element =
+            protocol_version_to_iana(result.client.actual_protocol_version().unwrap()).unwrap();
+        let slot = expected_version_element.slot_from_key().unwrap();
+        assert_eq!(record.negotiated_protocols.slots()[slot], 1);
+
+        let expected_cipher = result.client.cipher_suite().unwrap().to_owned();
+        let expected_cipher_element = Cipher::from_openssl_name(expected_cipher.as_str()).unwrap();
+        let slot = expected_cipher_element.slot_from_key().unwrap();
+        assert_eq!(record.negotiated_ciphers.slots()[slot], 1);
 
         let expected_group = result
             .client
             .selected_key_exchange_group()
             .unwrap()
             .to_owned();
-        assert_eq!(
-            record.negotiated_groups.count_for(expected_group.as_str()),
-            1
-        );
+        let expected_group_element: Group = expected_group.parse().unwrap();
+        let slot = expected_group_element.slot_from_key().unwrap();
+        assert_eq!(record.negotiated_groups.slots()[slot], 1);
 
         let expected_sig = result.client.signature_scheme().unwrap().to_owned();
-        assert_eq!(
-            record
-                .negotiated_signatures
-                .count_for(expected_sig.as_str()),
-            1
-        );
+        let expected_sig_element: Signature = expected_sig.parse().unwrap();
+        let slot = expected_sig_element.slot_from_key().unwrap();
+        assert_eq!(record.negotiated_signatures.slots()[slot], 1);
     }
 
     #[test]
@@ -545,10 +330,8 @@ mod tests {
         let _ = endpoint.client_handshake(&ARBITRARY_POLICY_1);
         endpoint.subscriber.finish_record();
         let records = endpoint.sink.records.lock().unwrap();
-        let record = &records[0].handshake;
+        let record = &records[0].as_schema().handshake;
 
-        /// For every slot in `counter`, assert the count is 1 iff the slot's
-        /// element description appears in `expected`, else 0.
         fn assert_supported_matches<const N: usize, T>(
             counter: &FrozenCounter<N, T>,
             expected: &[&str],
@@ -566,7 +349,7 @@ mod tests {
                 })
                 .collect();
 
-            for (slot, &count) in counter.slots_for_test().iter().enumerate() {
+            for (slot, &count) in counter.slots().iter().enumerate() {
                 let name = T::key_from_slot(slot).unwrap();
                 if expected_slots.contains(&slot) {
                     assert_eq!(count, 1, "{name} count is {count}, not one");
@@ -592,32 +375,55 @@ mod tests {
 
         endpoint.subscriber.finish_record();
         let records = endpoint.sink.records.lock().unwrap();
-        let record = &records[0].handshake;
+        let record = &records[0].as_schema().handshake;
 
         assert_eq!(record.handshake_count, 3);
-        assert_eq!(record.negotiated_ciphers.total(), 3);
-        assert_eq!(record.negotiated_groups.total(), 3);
-        assert_eq!(record.negotiated_signatures.total(), 3);
-        assert_eq!(record.negotiated_protocols.total(), 3);
+        assert_eq!(
+            record
+                .negotiated_ciphers
+                .iter_non_zero()
+                .map(|(_, _, c)| c)
+                .sum::<u64>(),
+            3
+        );
+        assert_eq!(
+            record
+                .negotiated_groups
+                .iter_non_zero()
+                .map(|(_, _, c)| c)
+                .sum::<u64>(),
+            3
+        );
+        assert_eq!(
+            record
+                .negotiated_signatures
+                .iter_non_zero()
+                .map(|(_, _, c)| c)
+                .sum::<u64>(),
+            3
+        );
+        assert_eq!(
+            record
+                .negotiated_protocols
+                .iter_non_zero()
+                .map(|(_, _, c)| c)
+                .sum::<u64>(),
+            3
+        );
     }
 
-    /// A record with no handshakes should be entirely empty/default.
     #[test]
     fn empty_record() {
         let endpoint = TestEndpoint::new();
 
         endpoint.subscriber.finish_record();
         let records = endpoint.sink.records.lock().unwrap();
-        let mut record = records[0].handshake.clone();
+        let mut record = records[0].as_schema().handshake.clone();
 
-        // ignore the freeze time, since that "default" value is set to the Unix Epoch.
         record.freeze_time = SystemTime::UNIX_EPOCH;
         assert_eq!(record, FrozenHandshakeRecord::default());
     }
 
-    /// ARBITRARY_POLICY_1 (20240503 / default_tls13) should be compatible with
-    /// General, Fips, and Cnsa1 profiles, but not CNSA2 (which requires MLKEM1024
-    /// and mldsa87).
     #[test]
     fn record_contents_compatibility_metrics() {
         let endpoint = TestEndpoint::new();
@@ -625,7 +431,7 @@ mod tests {
         endpoint.client_handshake(&ARBITRARY_POLICY_1);
         endpoint.subscriber.finish_record();
         let records = endpoint.sink.records.lock().unwrap();
-        let record = &records[0].handshake;
+        let record = &records[0].as_schema().handshake;
 
         assert_eq!(record.compatibility_general20251201, 1);
         assert_eq!(record.compatibility_fips20251201, 1);
@@ -633,10 +439,6 @@ mod tests {
         assert_eq!(record.compatibility_cnsa2, 0);
     }
 
-    /// Make sure that the compute time is less than the overall handshake time.
-    ///
-    /// Additionally, make sure that three handshakes takes longer than one handshake.
-    /// This provides some confidence that we are correctly e.g. adding amounts
     #[test]
     fn timers() {
         let endpoint = TestEndpoint::new();
@@ -644,7 +446,7 @@ mod tests {
         endpoint.client_handshake(&ARBITRARY_POLICY_1);
         endpoint.subscriber.finish_record();
         let records = endpoint.sink.records.lock().unwrap();
-        let single_handshake = &records[0].handshake;
+        let single_handshake = &records[0].as_schema().handshake;
 
         assert!(single_handshake.handshake_compute_us <= single_handshake.handshake_duration_us);
         drop(records);
@@ -654,8 +456,8 @@ mod tests {
         endpoint.client_handshake(&ARBITRARY_POLICY_1);
         endpoint.subscriber.finish_record();
         let records = endpoint.sink.records.lock().unwrap();
-        let single_handshake = &records[0].handshake;
-        let multiple_handshakes = &records[1].handshake;
+        let single_handshake = &records[0].as_schema().handshake;
+        let multiple_handshakes = &records[1].as_schema().handshake;
 
         assert!(
             multiple_handshakes.handshake_compute_us <= multiple_handshakes.handshake_duration_us
@@ -663,38 +465,5 @@ mod tests {
 
         assert!(single_handshake.handshake_compute_us < multiple_handshakes.handshake_compute_us);
         assert!(single_handshake.handshake_duration_us < multiple_handshakes.handshake_duration_us);
-    }
-
-    /// A JSON payload that only includes `handshake_count` deserializes
-    /// successfully: `#[serde(default)]` on every other field fills it with
-    /// its documented default (0 for integer counters, `SystemTime::UNIX_EPOCH`
-    /// for `freeze_time`, a zero-filled `FrozenCounter` for per-kind fields).
-    #[test]
-    fn deserialize_missing_fields_uses_defaults() {
-        let json = r#"{"handshake_count": 10}"#;
-        let record: FrozenHandshakeRecord = serde_json::from_str(json).unwrap();
-
-        assert_eq!(record.handshake_count, 10);
-        assert_eq!(record.freeze_time, SystemTime::UNIX_EPOCH);
-
-        // Per-kind counters default to a zero-filled slab of the right length.
-        assert_eq!(record.negotiated_protocols, FrozenCounter::default());
-        assert_eq!(record.negotiated_ciphers, FrozenCounter::default());
-        assert_eq!(record.negotiated_groups, FrozenCounter::default());
-        assert_eq!(record.negotiated_signatures, FrozenCounter::default());
-        assert_eq!(record.supported_protocols, FrozenCounter::default());
-        assert_eq!(record.supported_ciphers, FrozenCounter::default());
-        assert_eq!(record.supported_groups, FrozenCounter::default());
-        assert_eq!(record.supported_signatures, FrozenCounter::default());
-
-        // Scalar integer fields default to zero.
-        assert_eq!(record.sslv2_client_hello, 0);
-        assert_eq!(record.compatibility_general20251201, 0);
-        assert_eq!(record.compatibility_fips20251201, 0);
-        assert_eq!(record.compatibility_cnsa1, 0);
-        assert_eq!(record.compatibility_cnsa2, 0);
-        assert_eq!(record.handshake_duration_us, 0);
-        assert_eq!(record.handshake_compute_us, 0);
-        assert_eq!(record.synthetic_traffic_count, 0);
     }
 }
