+++ bindings/rust/standard/s2n-tls-metrics-subscriber/src/subscriber.rs
@@ -1,11 +1,11 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
+use s2n_tls_metrics_schema::record::FrozenHandshakeRecord;
+
 use crate::{
-    attribution::Attribution,
-    detector::SyntheticTrafficDetector,
-    record::{FrozenHandshakeRecord, HandshakeRecordInProgress, MetricRecord},
-    telemetry_sink::TelemetrySink,
+    Attribution, MetricRecord, detector::SyntheticTrafficDetector,
+    record::HandshakeRecordInProgress, telemetry_sink::TelemetrySink,
 };
 use arc_swap::ArcSwap;
 use s2n_tls::events::EventSubscriber;
@@ -155,7 +155,10 @@ impl<S: TelemetrySink> AggregatedMetricsSubscriber<S> {
 
         // This will block the thread until the record is received.
         let handshake = export_pipeline.metric_receiver.recv().unwrap();
-        let record = MetricRecord::new(handshake, self.inner.attribution.clone());
+        let record = MetricRecord::new(s2n_tls_metrics_schema::record::MetricRecord {
+            attribution: self.inner.attribution.clone().into_schema(),
+            handshake,
+        });
         export_pipeline.sink.export_record(&record);
         self.inner
             .last_export_epoch_ms
@@ -291,9 +294,9 @@ mod tests {
         );
 
         // Verify handshake counts
-        assert_eq!(records[0].handshake.handshake_count, 2);
-        assert_eq!(records[1].handshake.handshake_count, 1);
-        assert_eq!(records[2].handshake.handshake_count, 0);
+        assert_eq!(records[0].as_schema().handshake.handshake_count, 2);
+        assert_eq!(records[1].as_schema().handshake.handshake_count, 1);
+        assert_eq!(records[2].as_schema().handshake.handshake_count, 0);
     }
 
     /// Passive export: when the interval has elapsed, the next handshake
@@ -355,8 +358,6 @@ mod tests {
             atomic::{AtomicBool, Ordering},
         };
 
-        // Toggleable detector: returns whatever its inner flag is set to.
-        // Lets us drive "synthetic" vs "real" handshakes from the test body.
         #[derive(Debug)]
         struct ToggleDetector(Arc<AtomicBool>);
         impl SyntheticTrafficDetector for ToggleDetector {
@@ -397,13 +398,10 @@ mod tests {
         subscriber.finish_record();
         let records = sink.records.lock().unwrap();
         assert_eq!(records.len(), 1);
-        let record = &records[0].handshake;
+        let record = &records[0].as_schema().handshake;
 
-        // 2 real handshakes counted; 3 synthetic flagged separately and
-        // excluded from every other counter.
         assert_eq!(record.handshake_count, 2);
         assert_eq!(record.synthetic_traffic_count, 3);
-        // Negotiated counters reflect only the 2 real handshakes.
         assert_eq!(record.negotiated_protocols.total(), 2);
         assert_eq!(record.negotiated_ciphers.total(), 2);
     }
@@ -418,7 +416,7 @@ mod tests {
         endpoint.subscriber.finish_record();
 
         let records = endpoint.sink.records.lock().unwrap();
-        assert_eq!(records[0].handshake.handshake_count, 2);
-        assert_eq!(records[0].handshake.synthetic_traffic_count, 0);
+        assert_eq!(records[0].as_schema().handshake.handshake_count, 2);
+        assert_eq!(records[0].as_schema().handshake.synthetic_traffic_count, 0);
     }
 }
