+++ bindings/rust/standard/s2n-tls-metrics-subscriber/src/telemetry_sink.rs
@@ -3,7 +3,7 @@
 
 use std::sync::Arc;
 
-use crate::record::MetricRecord;
+use crate::MetricRecord;
 
 /// Trait abstracting the export destination for metric records.
 ///
@@ -12,19 +12,19 @@ use crate::record::MetricRecord;
 /// and writing it to a destination such as stdout, a file, or a network
 /// socket.
 ///
-/// `MetricRecord` implements both `serde::Serialize` and
-/// `metrique_writer::Entry`, so sinks can choose between serde-based
-/// formats (JSON via `serde_json`, CBOR via `ciborium`, etc.) or the
-/// metrique writer pipeline.
+/// `MetricRecord` implements `serde::Serialize` and `metrique_writer::Entry`,
+/// so sinks can choose between serde-based formats (JSON via `serde_json`,
+/// CBOR via `ciborium`, etc.) or the metrique writer pipeline.
+///
+/// For direct field access to the underlying schema types, use
+/// [`MetricRecord::as_schema()`] — but note that those types live in
+/// `s2n-tls-metrics-schema` which has no stability guarantees.
 pub trait TelemetrySink: Send + Sync + 'static {
     /// Export a single metric record.
     fn export_record(&self, record: &MetricRecord);
 }
 
 /// Blanket impl so that callers can pass an `Arc<T>` directly as a sink.
-/// This is useful when the caller wants to share a single sink instance
-/// across multiple subscribers — they wrap it in an `Arc` and pass clones
-/// without the underlying sink type needing to implement `Clone`.
 impl<T: TelemetrySink> TelemetrySink for Arc<T> {
     fn export_record(&self, record: &MetricRecord) {
         (**self).export_record(record)
