+++ bindings/rust/standard/s2n-tls-metrics-subscriber/src/test_utils.rs
@@ -9,8 +9,7 @@ use s2n_tls::{
 };
 
 use crate::{
-    AggregatedMetricsSubscriber, MetricRecord, attribution::Attribution,
-    telemetry_sink::TelemetrySink,
+    AggregatedMetricsSubscriber, Attribution, MetricRecord, telemetry_sink::TelemetrySink,
 };
 
 pub(crate) static ARBITRARY_POLICY_1: LazyLock<Policy> =
